feat: handle complex types
 - generates types in a single file instead of splitting types across files

closes #208
