--- example/output/ts/Category.ts
@@ -1,14 +1,0 @@
-export const Category = {
-  type: "object",
-  properties: {
-    id: {
-      type: "integer",
-      format: "int64",
-      minimum: -9223372036854776000,
-      maximum: 9223372036854776000,
-    },
-    name: { type: "string" },
-  },
-  title: "Category",
-  $id: "Category.json",
-} as const;
