--- example/output/ts/Customer.ts
@@ -1,28 +1,0 @@
-export const Customer = {
-  type: "object",
-  properties: {
-    id: {
-      type: "integer",
-      format: "int64",
-      minimum: -9223372036854776000,
-      maximum: 9223372036854776000,
-    },
-    username: { type: "string" },
-    address: {
-      type: "array",
-      items: {
-        type: "object",
-        properties: {
-          street: { type: "string" },
-          city: { type: "string" },
-          state: { type: "string" },
-          zip: { type: "string" },
-        },
-        title: "Address",
-        $id: "Address.json",
-      },
-    },
-  },
-  title: "Customer",
-  $id: "Customer.json",
-} as const;
