--- example/output/ts/Tag.ts
@@ -1,14 +1,0 @@
-export const Tag = {
-  type: "object",
-  properties: {
-    id: {
-      type: "integer",
-      format: "int64",
-      minimum: -9223372036854776000,
-      maximum: 9223372036854776000,
-    },
-    name: { type: "string" },
-  },
-  title: "Tag",
-  $id: "Tag.json",
-} as const;
