+++ example/output/ts/types.ts
@@ -1,0 +1,203 @@
+export const Order = {
+  type: "object",
+  properties: {
+    id: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    petId: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    quantity: {
+      type: "integer",
+      format: "int32",
+      minimum: -2147483648,
+      maximum: 2147483647,
+    },
+    shipDate: { type: "string", format: "date-time" },
+    status: {
+      type: "string",
+      description: "Order Status",
+      enum: ["placed", "approved", "delivered"],
+    },
+    complete: { type: "boolean" },
+  },
+  title: "Order",
+  $id: "Order.json",
+} as const;
+
+export const Customer = {
+  type: "object",
+  properties: {
+    id: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    username: { type: "string" },
+    address: {
+      type: "array",
+      items: {
+        type: "object",
+        properties: {
+          street: { type: "string" },
+          city: { type: "string" },
+          state: { type: "string" },
+          zip: { type: "string" },
+        },
+        title: "Address",
+        $id: "Address.json",
+      },
+    },
+  },
+  title: "Customer",
+  $id: "Customer.json",
+} as const;
+
+export const Address = {
+  type: "object",
+  properties: {
+    street: { type: "string" },
+    city: { type: "string" },
+    state: { type: "string" },
+    zip: { type: "string" },
+  },
+  title: "Address",
+  $id: "Address.json",
+} as const;
+
+export const Category = {
+  type: "object",
+  properties: {
+    id: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    name: { type: "string" },
+  },
+  title: "Category",
+  $id: "Category.json",
+} as const;
+
+export const User = {
+  type: "object",
+  properties: {
+    id: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    username: { type: "string" },
+    firstName: { type: "string" },
+    lastName: { type: "string" },
+    email: { type: "string" },
+    password: { type: "string" },
+    phone: { type: "string" },
+    userStatus: {
+      type: "integer",
+      description: "User Status",
+      format: "int32",
+      minimum: -2147483648,
+      maximum: 2147483647,
+    },
+  },
+  title: "User",
+  $id: "User.json",
+} as const;
+
+export const Tag = {
+  type: "object",
+  properties: {
+    id: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    name: { type: "string" },
+  },
+  title: "Tag",
+  $id: "Tag.json",
+} as const;
+
+export const Pet = {
+  required: ["name", "photoUrls"],
+  type: "object",
+  properties: {
+    id: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    name: { type: "string" },
+    category: {
+      type: "object",
+      properties: {
+        id: {
+          type: "integer",
+          format: "int64",
+          minimum: -9223372036854776000,
+          maximum: 9223372036854776000,
+        },
+        name: { type: "string" },
+      },
+      title: "Category",
+      $id: "Category.json",
+    },
+    photoUrls: { type: "array", items: { type: "string" } },
+    tags: {
+      type: "array",
+      items: {
+        type: "object",
+        properties: {
+          id: {
+            type: "integer",
+            format: "int64",
+            minimum: -9223372036854776000,
+            maximum: 9223372036854776000,
+          },
+          name: { type: "string" },
+        },
+        title: "Tag",
+        $id: "Tag.json",
+      },
+    },
+    status: {
+      type: "string",
+      description: "pet status in the store",
+      enum: ["available", "pending", "sold"],
+    },
+    nullableValue: {
+      type: ["string", "null"],
+      description: "example nullable value",
+    },
+  },
+  title: "Pet",
+  $id: "Pet.json",
+} as const;
+
+export const ApiResponse = {
+  type: "object",
+  properties: {
+    code: {
+      type: "integer",
+      format: "int32",
+      minimum: -2147483648,
+      maximum: 2147483647,
+    },
+    type: { type: "string" },
+    message: { type: "string" },
+  },
+  title: "ApiResponse",
+  $id: "ApiResponse.json",
+} as const;
