--- example/output/types/Address.d.ts
@@ -1,14 +1,0 @@
-/* eslint-disable */
-/**
- * This file was automatically generated by openapi-transformer-toolkit CLI/methods.
- * DO NOT MODIFY IT BY HAND. Instead, modify the source OpenAPI file,
- * and run openapi-transformer-toolkit CLI/methods to regenerate this file.
- */
-
-export interface Address {
-  street?: string;
-  city?: string;
-  state?: string;
-  zip?: string;
-  [k: string]: unknown;
-}
