--- example/output/types/Customer.d.ts
@@ -1,15 +1,0 @@
-import { Address } from './Address'
-
-/* eslint-disable */
-/**
- * This file was automatically generated by openapi-transformer-toolkit CLI/methods.
- * DO NOT MODIFY IT BY HAND. Instead, modify the source OpenAPI file,
- * and run openapi-transformer-toolkit CLI/methods to regenerate this file.
- */
-
-export interface Customer {
-  id?: number;
-  username?: string;
-  address?: Address[];
-  [k: string]: unknown;
-}
