--- example/output/types/Order.d.ts
@@ -1,19 +1,0 @@
-/* eslint-disable */
-/**
- * This file was automatically generated by openapi-transformer-toolkit CLI/methods.
- * DO NOT MODIFY IT BY HAND. Instead, modify the source OpenAPI file,
- * and run openapi-transformer-toolkit CLI/methods to regenerate this file.
- */
-
-export interface Order {
-  id?: number;
-  petId?: number;
-  quantity?: number;
-  shipDate?: string;
-  /**
-   * Order Status
-   */
-  status?: 'placed' | 'approved' | 'delivered';
-  complete?: boolean;
-  [k: string]: unknown;
-}
