--- example/output/types/Tag.d.ts
@@ -1,12 +1,0 @@
-/* eslint-disable */
-/**
- * This file was automatically generated by openapi-transformer-toolkit CLI/methods.
- * DO NOT MODIFY IT BY HAND. Instead, modify the source OpenAPI file,
- * and run openapi-transformer-toolkit CLI/methods to regenerate this file.
- */
-
-export interface Tag {
-  id?: number;
-  name?: string;
-  [k: string]: unknown;
-}
