--- example/output/types/User.d.ts
@@ -1,21 +1,0 @@
-/* eslint-disable */
-/**
- * This file was automatically generated by openapi-transformer-toolkit CLI/methods.
- * DO NOT MODIFY IT BY HAND. Instead, modify the source OpenAPI file,
- * and run openapi-transformer-toolkit CLI/methods to regenerate this file.
- */
-
-export interface User {
-  id?: number;
-  username?: string;
-  firstName?: string;
-  lastName?: string;
-  email?: string;
-  password?: string;
-  phone?: string;
-  /**
-   * User Status
-   */
-  userStatus?: number;
-  [k: string]: unknown;
-}
