+++ example/output/types/types.d.ts
@@ -1,27 +1,4 @@
-import tap from 'tap'
-import fs from 'fs-extra'
-import { runCommand } from '../src/commands/json2ts.js'
-import { resolveFromPackageRoot } from '../src/utils/paths.js'
-
-const TEST_DIRECTORY = resolveFromPackageRoot('test', 'temp')
-
-const inputPath = './test/fixtures/schemas'
-const outputPath = './test/temp/types-from-json'
-
-tap.test('json2ts runCommand', async t => {
-  fs.ensureDirSync(TEST_DIRECTORY)
-
-  await runCommand(inputPath, outputPath)
-
-  const generatedFiles = fs.readdirSync(outputPath)
-  t.match(generatedFiles, ['types.d.ts'], 'generates the expected TS files')
-
-  const TypeFile = resolveFromPackageRoot(outputPath, 'types.d.ts')
-  const generatedTypesFile = fs.readFileSync(TypeFile, 'utf-8')
-
-  t.same(
-    generatedTypesFile,
-    `/* eslint-disable */
+/* eslint-disable */
 /**
  * This file was automatically generated by openapi-transformer-toolkit CLI/methods.
  * DO NOT MODIFY IT BY HAND. Instead, modify the source OpenAPI file,
@@ -64,7 +41,7 @@ export interface Order {
   /**
    * Order Status
    */
-  status?: "placed" | "approved" | "delivered";
+  status?: 'placed' | 'approved' | 'delivered';
   complete?: boolean;
   [k: string]: unknown;
 }
@@ -78,7 +55,11 @@ export interface Pet {
   /**
    * pet status in the store
    */
-  status?: "available" | "pending" | "sold";
+  status?: 'available' | 'pending' | 'sold';
+  /**
+   * example nullable value
+   */
+  nullableValue?: string | null;
   [k: string]: unknown;
 }
 
@@ -102,7 +83,3 @@ export interface User {
   userStatus?: number;
   [k: string]: unknown;
 }
-`,
-    'types.d.ts is created correctly'
-  )
-})
