+++ test/oas2ts.test.js
@@ -15,35 +15,60 @@ tap.test('oas2ts', async t => {
     await runCommand(inputPath, outputPath)
 
     const generatedFiles = fs.readdirSync(outputPath)
-    t.match(
-      generatedFiles,
-      [
-        'Address.d.ts',
-        'ApiResponse.d.ts',
-        'Category.d.ts',
-        'Customer.d.ts',
-        'Order.d.ts',
-        'Pet.d.ts',
-        'Tag.d.ts',
-        'User.d.ts'
-      ],
-      'generates the expected TS files'
-    )
+    t.match(generatedFiles, ['types.d.ts'], 'generates the expected TS files')
 
-    const PetFile = resolveFromPackageRoot(outputPath, 'Pet.d.ts')
-    const generatedPetFile = fs.readFileSync(PetFile, 'utf-8')
+    const TypesFile = resolveFromPackageRoot(outputPath, 'types.d.ts')
+    const generatedTypesFile = fs.readFileSync(TypesFile, 'utf-8')
 
     t.same(
-      generatedPetFile,
-      `import { Category } from './Category'
-import { Tag } from './Tag'
-
-/* eslint-disable */
+      generatedTypesFile,
+      `/* eslint-disable */
 /**
  * This file was automatically generated by openapi-transformer-toolkit CLI/methods.
  * DO NOT MODIFY IT BY HAND. Instead, modify the source OpenAPI file,
  * and run openapi-transformer-toolkit CLI/methods to regenerate this file.
  */
+
+export interface Address {
+  street?: string;
+  city?: string;
+  state?: string;
+  zip?: string;
+  [k: string]: unknown;
+}
+
+export interface ApiResponse {
+  code?: number;
+  type?: string;
+  message?: string;
+  [k: string]: unknown;
+}
+
+export interface Category {
+  id?: number;
+  name?: string;
+  [k: string]: unknown;
+}
+
+export interface Customer {
+  id?: number;
+  username?: string;
+  address?: Address[];
+  [k: string]: unknown;
+}
+
+export interface Order {
+  id?: number;
+  petId?: number;
+  quantity?: number;
+  shipDate?: string;
+  /**
+   * Order Status
+   */
+  status?: "placed" | "approved" | "delivered";
+  complete?: boolean;
+  [k: string]: unknown;
+}
 
 export interface Pet {
   id?: number;
@@ -61,32 +86,29 @@ export interface Pet {
   nullableValue?: string | null;
   [k: string]: unknown;
 }
-`,
-      'Pet.d.ts is created correctly'
-    )
 
-    const CustomerFile = resolveFromPackageRoot(outputPath, 'Customer.d.ts')
-    const generatedCustomerFile = fs.readFileSync(CustomerFile, 'utf-8')
+export interface Tag {
+  id?: number;
+  name?: string;
+  [k: string]: unknown;
+}
 
-    t.same(
-      generatedCustomerFile,
-      `import { Address } from './Address'
-
-/* eslint-disable */
-/**
- * This file was automatically generated by openapi-transformer-toolkit CLI/methods.
- * DO NOT MODIFY IT BY HAND. Instead, modify the source OpenAPI file,
- * and run openapi-transformer-toolkit CLI/methods to regenerate this file.
- */
-
-export interface Customer {
+export interface User {
   id?: number;
   username?: string;
-  address?: Address[];
+  firstName?: string;
+  lastName?: string;
+  email?: string;
+  password?: string;
+  phone?: string;
+  /**
+   * User Status
+   */
+  userStatus?: number;
   [k: string]: unknown;
 }
 `,
-      'Customer.d.ts is created correctly'
+      'types.d.ts is created correctly'
     )
   })
 })
