+++ test/oas2tson.test.js
@@ -16,27 +16,144 @@ tap.test('oas2tson', async t => {
 
     const generatedFiles = fs.readdirSync(outputPath)
 
-    t.match(
-      generatedFiles,
-      [
-        'Address.ts',
-        'ApiResponse.ts',
-        'Category.ts',
-        'Customer.ts',
-        'Order.ts',
-        'Pet.ts',
-        'Tag.ts',
-        'User.ts'
-      ],
-      'generates the expected TS files'
-    )
+    t.match(generatedFiles, ['types.ts'], 'generates the expected TS files')
 
-    const PetFile = resolveFromPackageRoot(outputPath, 'Pet.ts')
-    const generatedPetFile = fs.readFileSync(PetFile, 'utf-8')
-
+    const TypesFile = resolveFromPackageRoot(outputPath, 'types.ts')
+    const generatedTypesFile = fs.readFileSync(TypesFile, 'utf-8')
     t.same(
-      generatedPetFile,
-      `export const Pet = {
+      generatedTypesFile,
+      `export const Order = {
+  type: "object",
+  properties: {
+    id: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    petId: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    quantity: {
+      type: "integer",
+      format: "int32",
+      minimum: -2147483648,
+      maximum: 2147483647,
+    },
+    shipDate: { type: "string", format: "date-time" },
+    status: {
+      type: "string",
+      description: "Order Status",
+      enum: ["placed", "approved", "delivered"],
+    },
+    complete: { type: "boolean" },
+  },
+  title: "Order",
+  $id: "Order.json",
+} as const;
+
+export const Customer = {
+  type: "object",
+  properties: {
+    id: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    username: { type: "string" },
+    address: {
+      type: "array",
+      items: {
+        type: "object",
+        properties: {
+          street: { type: "string" },
+          city: { type: "string" },
+          state: { type: "string" },
+          zip: { type: "string" },
+        },
+        title: "Address",
+        $id: "Address.json",
+      },
+    },
+  },
+  title: "Customer",
+  $id: "Customer.json",
+} as const;
+
+export const Address = {
+  type: "object",
+  properties: {
+    street: { type: "string" },
+    city: { type: "string" },
+    state: { type: "string" },
+    zip: { type: "string" },
+  },
+  title: "Address",
+  $id: "Address.json",
+} as const;
+
+export const Category = {
+  type: "object",
+  properties: {
+    id: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    name: { type: "string" },
+  },
+  title: "Category",
+  $id: "Category.json",
+} as const;
+
+export const User = {
+  type: "object",
+  properties: {
+    id: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    username: { type: "string" },
+    firstName: { type: "string" },
+    lastName: { type: "string" },
+    email: { type: "string" },
+    password: { type: "string" },
+    phone: { type: "string" },
+    userStatus: {
+      type: "integer",
+      description: "User Status",
+      format: "int32",
+      minimum: -2147483648,
+      maximum: 2147483647,
+    },
+  },
+  title: "User",
+  $id: "User.json",
+} as const;
+
+export const Tag = {
+  type: "object",
+  properties: {
+    id: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    name: { type: "string" },
+  },
+  title: "Tag",
+  $id: "Tag.json",
+} as const;
+
+export const Pet = {
   required: ["name", "photoUrls"],
   type: "object",
   properties: {
@@ -92,16 +209,76 @@ tap.test('oas2tson', async t => {
   title: "Pet",
   $id: "Pet.json",
 } as const;
+
+export const ApiResponse = {
+  type: "object",
+  properties: {
+    code: {
+      type: "integer",
+      format: "int32",
+      minimum: -2147483648,
+      maximum: 2147483647,
+    },
+    type: { type: "string" },
+    message: { type: "string" },
+  },
+  title: "ApiResponse",
+  $id: "ApiResponse.json",
+} as const;
 `,
-      'Pet.ts is created correctly'
+      'types.ts is created correctly'
     )
+  })
 
-    const CustomerFile = resolveFromPackageRoot(outputPath, 'Customer.ts')
-    const generatedCustomerFile = fs.readFileSync(CustomerFile, 'utf-8')
+  t.test('runCommand function with excludeDereferencedIds', async t => {
+    fs.ensureDirSync(TEST_DIRECTORY)
 
+    await runCommand(inputPath, outputPath, null, true)
+
+    const generatedFiles = fs.readdirSync(outputPath)
+
+    t.match(generatedFiles, ['types.ts'], 'generates the expected TS files')
+
+    const TypesFile = resolveFromPackageRoot(outputPath, 'types.ts')
+    const generatedTypesFile = fs.readFileSync(TypesFile, 'utf-8')
+    // console.log('************************************************************')
+    // console.log(generatedTypesFile)
+    // console.log('************************************************************')
     t.same(
-      generatedCustomerFile,
-      `export const Customer = {
+      `export const Order = {
+  type: "object",
+  properties: {
+    id: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    petId: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    quantity: {
+      type: "integer",
+      format: "int32",
+      minimum: -2147483648,
+      maximum: 2147483647,
+    },
+    shipDate: { type: "string", format: "date-time" },
+    status: {
+      type: "string",
+      description: "Order Status",
+      enum: ["placed", "approved", "delivered"],
+    },
+    complete: { type: "boolean" },
+  },
+  title: "Order",
+  $id: "Order.json",
+} as const;
+
+export const Customer = {
   type: "object",
   properties: {
     id: {
@@ -122,46 +299,83 @@ tap.test('oas2tson', async t => {
           zip: { type: "string" },
         },
         title: "Address",
-        $id: "Address.json",
       },
     },
   },
   title: "Customer",
   $id: "Customer.json",
 } as const;
-`,
-      'Customer.ts is created correctly'
-    )
-  })
 
-  t.test('runCommand function with excludeDereferencedIds', async t => {
-    fs.ensureDirSync(TEST_DIRECTORY)
+export const Address = {
+  type: "object",
+  properties: {
+    street: { type: "string" },
+    city: { type: "string" },
+    state: { type: "string" },
+    zip: { type: "string" },
+  },
+  title: "Address",
+  $id: "Address.json",
+} as const;
 
-    await runCommand(inputPath, outputPath, null, true)
+export const Category = {
+  type: "object",
+  properties: {
+    id: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    name: { type: "string" },
+  },
+  title: "Category",
+  $id: "Category.json",
+} as const;
 
-    const generatedFiles = fs.readdirSync(outputPath)
+export const User = {
+  type: "object",
+  properties: {
+    id: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    username: { type: "string" },
+    firstName: { type: "string" },
+    lastName: { type: "string" },
+    email: { type: "string" },
+    password: { type: "string" },
+    phone: { type: "string" },
+    userStatus: {
+      type: "integer",
+      description: "User Status",
+      format: "int32",
+      minimum: -2147483648,
+      maximum: 2147483647,
+    },
+  },
+  title: "User",
+  $id: "User.json",
+} as const;
 
-    t.match(
-      generatedFiles,
-      [
-        'Address.ts',
-        'ApiResponse.ts',
-        'Category.ts',
-        'Customer.ts',
-        'Order.ts',
-        'Pet.ts',
-        'Tag.ts',
-        'User.ts'
-      ],
-      'generates the expected TS files'
-    )
+export const Tag = {
+  type: "object",
+  properties: {
+    id: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    name: { type: "string" },
+  },
+  title: "Tag",
+  $id: "Tag.json",
+} as const;
 
-    const PetFile = resolveFromPackageRoot(outputPath, 'Pet.ts')
-    const generatedPetFile = fs.readFileSync(PetFile, 'utf-8')
-
-    t.same(
-      generatedPetFile,
-      `export const Pet = {
+export const Pet = {
   required: ["name", "photoUrls"],
   type: "object",
   properties: {
@@ -215,16 +429,57 @@ tap.test('oas2tson', async t => {
   title: "Pet",
   $id: "Pet.json",
 } as const;
-`,
-      'Pet.ts is created correctly'
-    )
 
-    const CustomerFile = resolveFromPackageRoot(outputPath, 'Customer.ts')
-    const generatedCustomerFile = fs.readFileSync(CustomerFile, 'utf-8')
+export const ApiResponse = {
+  type: "object",
+  properties: {
+    code: {
+      type: "integer",
+      format: "int32",
+      minimum: -2147483648,
+      maximum: 2147483647,
+    },
+    type: { type: "string" },
+    message: { type: "string" },
+  },
+  title: "ApiResponse",
+  $id: "ApiResponse.json",
+} as const;
 
-    t.same(
-      generatedCustomerFile,
-      `export const Customer = {
+export const Order = {
+  type: "object",
+  properties: {
+    id: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    petId: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    quantity: {
+      type: "integer",
+      format: "int32",
+      minimum: -2147483648,
+      maximum: 2147483647,
+    },
+    shipDate: { type: "string", format: "date-time" },
+    status: {
+      type: "string",
+      description: "Order Status",
+      enum: ["placed", "approved", "delivered"],
+    },
+    complete: { type: "boolean" },
+  },
+  title: "Order",
+  $id: "Order.json",
+} as const;
+
+export const Customer = {
   type: "object",
   properties: {
     id: {
@@ -251,8 +506,149 @@ tap.test('oas2tson', async t => {
   title: "Customer",
   $id: "Customer.json",
 } as const;
+
+export const Address = {
+  type: "object",
+  properties: {
+    street: { type: "string" },
+    city: { type: "string" },
+    state: { type: "string" },
+    zip: { type: "string" },
+  },
+  title: "Address",
+  $id: "Address.json",
+} as const;
+
+export const Category = {
+  type: "object",
+  properties: {
+    id: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    name: { type: "string" },
+  },
+  title: "Category",
+  $id: "Category.json",
+} as const;
+
+export const User = {
+  type: "object",
+  properties: {
+    id: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    username: { type: "string" },
+    firstName: { type: "string" },
+    lastName: { type: "string" },
+    email: { type: "string" },
+    password: { type: "string" },
+    phone: { type: "string" },
+    userStatus: {
+      type: "integer",
+      description: "User Status",
+      format: "int32",
+      minimum: -2147483648,
+      maximum: 2147483647,
+    },
+  },
+  title: "User",
+  $id: "User.json",
+} as const;
+
+export const Tag = {
+  type: "object",
+  properties: {
+    id: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    name: { type: "string" },
+  },
+  title: "Tag",
+  $id: "Tag.json",
+} as const;
+
+export const Pet = {
+  required: ["name", "photoUrls"],
+  type: "object",
+  properties: {
+    id: {
+      type: "integer",
+      format: "int64",
+      minimum: -9223372036854776000,
+      maximum: 9223372036854776000,
+    },
+    name: { type: "string" },
+    category: {
+      type: "object",
+      properties: {
+        id: {
+          type: "integer",
+          format: "int64",
+          minimum: -9223372036854776000,
+          maximum: 9223372036854776000,
+        },
+        name: { type: "string" },
+      },
+      title: "Category",
+    },
+    photoUrls: { type: "array", items: { type: "string" } },
+    tags: {
+      type: "array",
+      items: {
+        type: "object",
+        properties: {
+          id: {
+            type: "integer",
+            format: "int64",
+            minimum: -9223372036854776000,
+            maximum: 9223372036854776000,
+          },
+          name: { type: "string" },
+        },
+        title: "Tag",
+      },
+    },
+    status: {
+      type: "string",
+      description: "pet status in the store",
+      enum: ["available", "pending", "sold"],
+    },
+    nullableValue: {
+      type: ["string", "null"],
+      description: "example nullable value",
+    },
+  },
+  title: "Pet",
+  $id: "Pet.json",
+} as const;
+
+export const ApiResponse = {
+  type: "object",
+  properties: {
+    code: {
+      type: "integer",
+      format: "int32",
+      minimum: -2147483648,
+      maximum: 2147483647,
+    },
+    type: { type: "string" },
+    message: { type: "string" },
+  },
+  title: "ApiResponse",
+  $id: "ApiResponse.json",
+} as const;
 `,
-      'Customer.ts is created correctly'
+      generatedTypesFile,
+      'types.ts is created correctly'
     )
   })
 })
